seeded data into roles table

+++ db/schema.sql
@@ -14,7 +14,7 @@ CREATE TABLE department (
 CREATE TABLE roles (
     id INT PRIMARY KEY AUTO_INCREMENT NOT NULL,
     title VARCHAR(30) NOT NULL,
-    salary DECIMAL NOT NULL,
+    salary INT NOT NULL,
     department_id INT NOT NULL,
     FOREIGN KEY (university_id)
         REFERENCES department(id)
